Fix types

+++ src/ACT31.cpp
@@ -9,7 +9,7 @@ using namespace std;
 * @return feet * 12
 *
 */
-int feetToInches(int feet) {
+double feetToInches(double feet) {
 	return (feet * 12);
 }
 
@@ -19,22 +19,23 @@ int feetToInches(int feet) {
 * @return feet * 30.5
 *
 */
-double feetToCM(int feet) {
+double feetToCM(double
+		feet) {
 	return (feet * 30.5);
 }
 
 int main() {
-	int feet;
+	double feet;
 	cout << "Please enter a number of feet: ";
 	cin >> feet;
 
-	int inches;
+	double inches;
 	inches = feetToInches(feet);
 
 	double centimeters;
 	centimeters = feetToCM(feet);
 
-	printf("\nThere are %d inches and %.1f centimeters in %d feet.",
+	printf("\nThere are %.2f inches and %.2f centimeters in %.2f feet.",
 			inches, centimeters, feet);
 
 	return 0;
